fix: pretty format

+++ contracts/strategies/StrategyCurve3CrvVoterProxy.sol
@@ -41,9 +41,9 @@ contract StrategyCurve3CrvVoterProxy {
     address public controller;
     address public strategist;
 
-    uint256 public earned;  // lifetime strategy earnings denominated in `want` token
+    uint256 public earned; // lifetime strategy earnings denominated in `want` token
 
-    event Harvested(uint wantEarned, uint lifetimeEarned);
+    event Harvested(uint256 wantEarned, uint256 lifetimeEarned);
 
     constructor(address _controller) public {
         governance = msg.sender;
@@ -75,7 +75,7 @@ contract StrategyCurve3CrvVoterProxy {
         performanceFee = _performanceFee;
     }
 
-    function setStrategistReward(uint _strategistReward) external {
+    function setStrategistReward(uint256 _strategistReward) external {
         require(msg.sender == governance, "!governance");
         strategistReward = _strategistReward;
     }
